pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer – это сайт, который анализирует указанные страницы на SEO-пригодность по аналогии с PageSpeed Insights"
authors = ["NevermoreKatana <grenkachannel@hotmail.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10.6"
Flask = "^2.3.3"
gunicorn = "^21.2.0"
flake8 = "^6.1.0"
validators = "^0.22.0"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

